refactor: export all hooks functions directly from main module
refactor: rename the modules to make sure tree-shaking works
