+++ components/index.ts
@@ -1,16 +1,21 @@
 /// <reference types="styled-jsx" />
 
 export * from './styles/themes'
-export { default as Utils } from './utils-shared/index'
+export { default as Utils } from './utils-shared-deprecated/index'
 export { default as useTheme } from './styles/use-theme'
 export { default as ZEITUIProvider } from './zeit-provider-deprecated'
 export { default as ZeitProvider } from './zeit-provider'
 export { default as CSSBaseline } from './css-baseline-deprecated'
 export { default as CssBaseline } from './css-baseline'
-export { default as useToasts } from './toast'
+export { default as useToasts } from './use-toasts'
 export { default as useInput } from './input/use-input'
 export { default as useModal } from './modal/use-modal'
 export { default as useTabs } from './tabs/use-tabs'
+export { default as useBodyScroll } from './use-body-scroll'
+export { default as useClickAway } from './use-click-away'
+export { default as useClipboard } from './use-clipboard'
+export { default as useCurrentState } from './use-current-state'
+export { default as useMediaQuery } from './use-media-query'
 export { default as Avatar } from './avatar'
 export { default as Text } from './text'
 export { default as Note } from './note'
@@ -39,7 +44,7 @@ export { default as Radio } from './radio'
 export { default as Select } from './select'
 export { default as Tabs } from './tabs'
 export { default as Progress } from './progress'
-export { default as Tree } from './file-tree'
+export { default as Tree } from './tree'
 export { default as Badge } from './badge'
 export { default as AutoComplete } from './auto-complete'
 export { default as Collapse } from './collapse'
