+++ components/tree/tree-folder-icon.tsx
