+++ components/tree/tree-status-icon.tsx
