+++ components/use-body-scroll/index.ts
@@ -1,0 +1,3 @@
+import useBodyScroll from './use-body-scroll'
+
+export default useBodyScroll
