+++ components/use-click-away/index.ts
@@ -1,0 +1,3 @@
+import useClickAway from './use-click-away'
+
+export default useClickAway
