+++ components/use-clipboard/index.ts
@@ -1,0 +1,3 @@
+import useClipboard from './use-clipboard'
+
+export default useClipboard
