+++ components/use-clipboard/use-clipboard.ts
