+++ components/use-current-state/index.ts
@@ -1,0 +1,3 @@
+import useCurrentState from './use-current-state'
+
+export default useCurrentState
