+++ components/use-media-query/index.ts
@@ -1,0 +1,3 @@
+import useMediaQuery from './use-media-query'
+
+export default useMediaQuery
