--- components/utils-shared/index.ts
@@ -1,13 +1,0 @@
-import { default as useBodyScroll } from './use-body-scroll'
-import { default as useClipboard } from './use-clipboard'
-import { default as useCurrentState } from './use-current-state'
-import { default as useClickAway } from './use-click-away'
-import { default as useMediaQuery } from './use-media-query'
-
-export default {
-  useBodyScroll,
-  useClipboard,
-  useCurrentState,
-  useClickAway,
-  useMediaQuery,
-}
