+++ components/utils/use-body-scroll.ts
@@ -1,3 +1,3 @@
-import useBodyScroll from '../utils-shared/use-body-scroll'
+import useBodyScroll from '../utils-shared-deprecated/use-body-scroll'
 
 export default useBodyScroll
