+++ components/utils/use-clipboard.ts
@@ -1,3 +1,3 @@
-import useClipboard from '../utils-shared/use-clipboard'
+import useClipboard from '../utils-shared-deprecated/use-clipboard'
 
 export default useClipboard
