+++ components/utils/use-current-state.ts
@@ -1,3 +1,3 @@
-import useCurrentState from '../utils-shared/use-current-state'
+import useCurrentState from '../utils-shared-deprecated/use-current-state'
 
 export default useCurrentState
